test/pylib: move palette color helpers from suite.base to util
Relocate the `palette` class, `create_formatter()` and the
`output_is_a_tty` flag from test/pylib/suite/base.py into
test/pylib/util.py, and update the importers (test.py and
test/pylib/cql_repl.py) accordingly.
